Tune SQLAlchemy pool parameters
We can burst and create a lot of threads, each of which will checkout
a SQLAlchemy connection from the pool.  This accomodates that.

We have a natural limit on the number of db connections -- we will
never use more than the total number of servers managed.  So in that
case, just don't set an overflow limit for the db connection pool.
This means that it will stabilize on 5 open connections and burst
to as many as needed.

Also, ensure that the connection is returned to the pool in the
context manager exit method, as well as setting it to None so that
the session can not be re-used again (this is an easy way to make
sure that it can't be used except as a context manager).

Change-Id: Ie4628326b6b84fb0979e4eceed546404c4e30637