Inspect the Gearman queue for immediate demand
Use information from Gearman to determine the immediate load
requirements of the system and spin up as many nodes as required
to meet the demand.  Use the existing information about the
min and max servers to determine the ratio of servers to spin
up from each provider.

Replaces the several fake server scripts with one script that
implements statsd, zmq, and gearman to ease testing.

Change-Id: Ic0dedc7ef2760ff664912f771377e02967ad5633