Default to a ratelimit of 2/second for API calls
We make a lot of API calls when deleting servers - 8 in fact to
enumerate and delete floating ips, keypairs and the server. A 1 second
rate limit between calls makes each delete take 8 seconds to move
past, and then if we're deleting 50 servers on a single provider we
won't get through submitting the deletes for all 50 for 500 seconds,
or 8 minutes.

Change-Id: I4ce86c38030e63db3547043134940fb9cf01eafa