Log the time a node has been in state DELETE.
Previously we logged the time spent in deleteNode which can be
dramatically different - for instance when the 10 minute timeout
expires, we don't delete the node, and when periodic cleanup does
delete the node it will be gone already so we may only spend seconds
doing the deletion. The total elapsed time is actually much more
interesting.

Change-Id: If5a4561262663fcf265c3060a2034901cd8a8c3c