Expose paramiko's get_pty parameter.
This is needed for Fedora - we could just force it on everywhere, or
someone could add glue to control it on a per-image basis. However
anyone testing commands interactively will have a tty, so forcing it
on should lead to less surprises. I've made the patch expose the
option so that making it more specific in future is simple.

Change-Id: I41173e06201421d5d77d23155d925fe060e478fc