Create new provider managers on image data changes
Previously only non image provider config would force nodepool to create
a new provider manager. There are cases when it is useful to  have
nodepool create a new manager when the image config changes. For
example, when you want to add an image or change the flavor of an image.

Add additional checks to see if the list of image names has changed and
if all the image names remain the same check if any of their properties
have changed. If any of these things have changed create a new provider
manager.

Change-Id: Ic0732453e12069d5d50a56dea0037817e978dad4