Don't accept tasks for stopped managers
In case a provider manager is being replaced due to a configuration
change, raise exceptions for any tasks submitted after it is stopped.

This will probably cause many errors, but they should already be
handled and servers will eventually be deleted.

We can minimize the disruption by making the provider managers more
adaptable to changes, but this stopgap measure should at least fix
the current problem we are observing with threads that get stuck
and never complete.

Change-Id: I3d190881ede30480d7c4ae970a0cb2dd07c3e160