Add support for network labels
This patch adds the ability to specify a net-label
instead of using a net-id (network UUID).

Rather than use network UUID's in our nodepool.yaml
config files it would be nice to use the more meaningful
network labels instead. This should make the config file
more readable across the various cloud providers and
matches how we use image names (instead of image
UUIDs) as well.

The current implementation relies on the os-tenant-networks
extension in Nova to provide the network label lookup.
Given that nodepool is currently focused around using
novaclient this made the most sense. We may at some point
in the future want to use the Neutron API directly for
this information or perhaps use a combination of both
approaches to accommodate a variety of provider API
deployment choices.

Tested locally on my TripleO overcloud using two
Neutron networks.

Change-Id: I9bdd35adf2d85659cf1b992ccd2fcf98fb124528