Add stdout and stderr to exception when ready script fails
When a command run through ssh fails, nodepool raises an exception
saying that it failed, but doesn't include any other information.
This patch adds in the output from stderr and stdout if output=True,
as well as the exact command run to help debug problems.

Also, added output parameter to the fake ssh client for testing.

Change-Id: If55643aef91c90b7c27fe4f532d51d9ef72b1ab4