Fix potential floating ip leakage issue
Previously, when cleanupServer() was called it would first query nova
to get the server and then release the floatingip associated with it
(if there was one) and then delete the floatingip. The potential floating
ip leak would occur if we called removeFloatingIP and the call to
deleteFloatingIP failed. The leak would occur the next time we tried to
clean up the server as we no longer find the floatingip to delete as
it had been disassocated from the server. To fix this issue we just need
to drop the call to removeFloating and call deleteFloatingIP as we don't
really need to disassociate the floatingip first.

Unfortunately, there is a potential bug in nova-api when using neutron that
this call can disassociate the floatingip and then fail to delete it. In this
case we'll end up leaking the floatingip in the same way. The fix for this
nova issue is here: I53b0c9d949404288e8687b304361a74b53d69ef3

closes-bug: 1356161

Change-Id: I0c78823198fac0d31235d93505a4251edbf9e612