Move pool.stop into addCleanup
If the test is failing because nodes don't come up correctly, tests
will be stuck in waitForNodes().  This will eventually hit the
test-timeout (60 seconds) where python-fixtures will raise
TimeoutException (because gentle=True is set in BaseTestCase).  This
kills the test function, but pool.stop() never gets called.

The threads created by Nodepool.Nodepool() aren't daemon threads, so
this causes testr to not exit and just sit-around waiting forever when
a test fails.

The easiest way to avoid this is to add a cleanup-handler that calls
pool.stop() after the pool has been created.  This works in both the
success and failure case

Change-Id: Icee97498938f12001d873bb78217c470a3fe03e7