Remove setUp() call from setup_config
class TestNodepool() subclasses tests.DBTestCase, which provides
setUp() ... but then when setup_config is called it will start the
setUp() call chain again.

One result of this is logging gets pretty messed up after a call to
setup_config().  There is also an open upstream bug to detect this
[1], so it might break in the future.

There doesn't seem to be any need to call it in setup_config()

[1] https://bugs.launchpad.net/testtools/+bug/882884

Change-Id: I6e646c70656348cee64c55ce83ab90fecdbaa8d4