Delete correct item in image delete command.
Previously the nodepool delete image command got a snapshot image object
from the database and passed that to nodepool.deleteImage(). Now
deleteImage() expects just the image id number and will populate the
image object from the database itself. Update the image delete command
to pass the id and not the image object.

Change-Id: Ibfe815903d2754b660e011fcdefe47aecfad6d91