Line-by-line dib logs
Log each line of output from DIB as it is read, and log it to the
image log file instead of the standard one.

Also, update the fake-dib system so that it exercises more code.
Fix the filename check for fake-dib as the fake filename key
actually ends up in the middle of the filename string.

Add a fake-dib config file.

Change-Id: I1e43c26ac0f9db4cf4e52f77b2cf5ca59f19989c