Change sequence on actions in updateConfig
There is an incorrect sequence of actions in updateConfig()
function. The problem is that in case if gearman server is not
available the log file will contain number of error messages:

    'NoneType' object has no attribute 'db'

This is because reconfigureGearmanClient() waits for the servers
and blocks call to setConfig() function. At the same time there
are number of asynchronous periodic check that have been started.
They try to access self.config.db via _geDB() function, which,
indeed, is None at that time.

So I propose to call reconfigureCrons last before calling setConfig().

Change-Id: Ic55d9e6bc17b10a688bf13c6a076f608073c7dc9