Add current node IP to /etc/nodepool
Record the IP of the current node in /etc/nodepool.  Also remove
the try/except that should no longer be necessary (consider lack
of /etc/nodepool an error now).  Add some comments about what the
files are.  Update docs to include this change and a prior change
that added the provider file.

Change-Id: I730e1ea7c67a9cd9c6bcd9989c4c24d7778cded9