Set useful PATH before starting scripts
In change I9d9f759fac1a96cc785354ae5aa44398c908704d I mentioned that
CentOS 7 rax images seemed to be working with the scripts, but hpcloud
images were failing.

What I didn't realise was the combination of two things -- firstly the
ssh session run by paramiko doesn't source anything gets sourced.  The
other difference is that we log into the rax images as 'root', while
the hpcloud images require logging in as 'centos'
(I276c44b2c711e1d8db980c8ee39a07d912cfc720).  So you'll fall back to
the in-built ssh paths, which are

 --with-default-path=/usr/local/bin:/usr/bin
 --with-superuser-path=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin

at least for centos.

The proposes to standard the PATH set for scripts so it doesn't matter
if they are running from a 'root' user or a 'cloud' user via sudo.

The hpcloud images are building, but there are still some errors in
there from some sbin tools.  It's quite reasonable for the scripts to
expect they're running with that superuser-path, so let's set it
before starting them.

Change-Id: Ic038b09582472856be604300680e2f2cc269f9cf