Caching floating ips as server list is done
There are individual requets for each node to be built and
deleted to query for floating ips list. That means one call
to the nova API for each process, causing high load into the
system. Adding a cache layer that only queries for the list
every 5 seconds, as well as we are doing for the server list.

Change-Id: I43dd4805debbdf790c3ec28db686459138d59d4d