Match novaclients transitive six dep
Because of the way pip does dep resolution the "highest" dep
specification wins. In this case the nodepool specified six dep trumps
the six dep that we have transitively via novaclient. Unfortunately this
means that if you have an old six already installed and novaclient needs
a newer version that newer version is not installed.

Fix this by moving nodepools six dep forward to match novaclient.

Change-Id: Ic6f3611bea8c47bdc40757d134670c5f0ed8dc13