Update snapshots prior to dib image builts
In the daily image updates, dib images were taking priority
over snapshots. As we were waiting for dib images to be built
prior to uploading, it was preventing the snapshot updates
to be started until all dib builds were finished. In this
change i update the snapshots first, then i build the dib
images and then i upload the resulting dib images. By this
way we are not blocking anything but image uploads until
all the dib images are processed.

Change-Id: Ifc2e205b71fe65bf2f7183632942ba52d62ad208