Fix error handling when dib image builds fail
Old code to build dib images would "succeed" even when image builds had
failed due to diskimage-create returning a non zero value. This was
because we caught and suppressed failures too close the failure so that
calling code was unaware and would make the builds as successful in the
DB and in the log.

Fix this by letting the errors bubble up to code that deals with logging
and commiting this data to the DB so that it can decide when things have
actually succeeded.

Story: 2000067
Change-Id: Ibe8209a2ecc55a8c7dfaf7887d0486079d20e27c