Fix nodepool --version
Currently, the `nodepool --version` command prints a usage message
followed by 'nodepool: error: too few arguments' because we are
not specifying a subcommand. The proper way to do this is to use the
version action in argparser rather than store_true with some logic.

Add version_info so it can successfully be imported and a version
action to print the imported string.

Change-Id: Ic5437d508b27fe4a8e02361cfe345c1f741d3c48