Cleanup DIB images in DB if they fail to build
When DIB fails to build an image for whatever reason remove that image
from the DB. Otherwise we have an image in the building state for
eternity. DB should clean up after itself on disk so we don't need to
worry about that and we will not have made it far enough to upload to
glance. The only thing that needs cleanup here is the database.

Change-Id: I34e643ade18c273b7fe74642786101e87d44699c