Pin python-daemon to <2.0
The 2.0 release series of python-daemon has been broken for a few
reasons. At one point it had an undeclared dependency on docutils that
caused things to break. Current issue is it breaks the egg_info writers
of other packages like extras and pyyaml.

Just pin to a known working release for now so that we can fix
nodepool's lack of timeouts.

Change-Id: I11c47c15fdc1a0d4bd6cd8e8ae6dfe88083e95e6