Iterate over images, not labels, when updating
Labels do not map 1:1 to images, so when nodepool was updating
images by iterating over labels, it may have updated an image
more than once.  This was causing duplicate snapshot and DIB
images (though not duplicate DIB files).

Instead, iterate over provider image entries (which are unique)
when deciding which to update.

This includes a small unrelated change to the fake dib selector
used in debugging the problem leading to this patch.

Change-Id: I4e03c1136376b51987e9f2f525ed6613ed4500eb